fix: out of range behavior is inconsistent with curve remapping and edit-mode behavior
